+++ packages/orq-rc/RELEASES.md
@@ -5319,3 +5319,13 @@ Based on:
 - [typescript v4.0.0-rc.66] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.66] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.66 - packages/orq-rc
+
+## 2025-11-22 10:51:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.660.0 (2.760.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.67] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.67] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.67 - packages/orq-rc